게시글 삭제시 권한 확인

+++ src/main/java/com/example/prj_practice_back/controller/BoardController.java
@@ -1,12 +1,15 @@
 package com.example.prj_practice_back.controller;
 
 import com.example.prj_practice_back.domain.Board;
+import com.example.prj_practice_back.domain.Member;
 import com.example.prj_practice_back.service.BoardService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Map;
 
 @RestController
 @RequiredArgsConstructor
@@ -16,11 +19,20 @@ public class BoardController {
     private final BoardService service;
 
     @PostMapping("add")
-    public ResponseEntity add(@RequestBody Board board) {
+    public ResponseEntity add(@RequestBody Board board,
+                              @SessionAttribute(value = "login", required = false) Member login) {
 
+        if (login == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
 
+        System.out.println("login = " + login);
 
-        if (service.save(board)) {
+        if (!service.validate(board)) {
+            return ResponseEntity.badRequest().build();
+        }
+
+        if (service.save(board,login)) {
             return ResponseEntity.ok().build();
         } else {
             return ResponseEntity.internalServerError().build();
@@ -38,7 +50,14 @@ public class BoardController {
     }
 
     @DeleteMapping("remove/{id}")
-    public ResponseEntity remove(@PathVariable Integer id) {
+    public ResponseEntity remove(@PathVariable Integer id,
+                                 @SessionAttribute (value = "login", required = false)Member login) {
+        if (login != null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build(); //401 code
+        }
+        if (!service.hasAccess(id, login)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build(); //403 code
+        }
         if (service.remove(id)) {
             return ResponseEntity.ok().build();
         } else {
@@ -48,9 +67,7 @@ public class BoardController {
 
     @PutMapping("edit")
     public ResponseEntity edit(@RequestBody Board board) {
-//        System.out.println("board = " + board);
         if (service.validate(board)) {
-
             if (service.update(board)) {
                 return ResponseEntity.ok().build();
             } else {
@@ -59,7 +76,5 @@ public class BoardController {
         } else {
             return ResponseEntity.badRequest().build();
         }
-
     }
-
 }
